Add ftag alias
The ftag alias is designed to sync the tags that exist on the remote
with the tags that exist in the local repository, which includes pruning
tags locally that no longer exist on the remote. Thus, it is
unidirectional in that it assumes that the tags on the remote are the
proper state of things.

c.f. https://stackoverflow.com/a/26282566